REST-server for desktop application
Implement the first draft of a REST server which can later be used to
communicate with a GUI or other CLI-clients for qabel-desktop.

Note: For testing the qabel-core project is needed in ../qabel-core
because this build.gradle extends the java library path to ../qabel-core
to include the binary library for crypto stuff.